refactor to function
